+++ README.md
@@ -17,5 +17,7 @@ A desktop application to rename photos and videos using project numbers, tags an
    ```
 
 Configuration files are stored in the user specific configuration directory
-(e.g. `~/.config/mic-renamer` on Linux). Set the `RENAMER_CONFIG_DIR`
-environment variable to override this location.
+(e.g. `~/.config/mic_renamer` on Linux). Set the `RENAMER_CONFIG_DIR`
+environment variable to override this location. Defaults are bundled in
+`mic_renamer/config/defaults.yaml` and will be merged with your custom
+configuration at startup.
