+++ mic_renamer/config/__init__.py
@@ -1,0 +1,1 @@
+from .config_manager import ConfigManager
