--- mic_renamer/config/app_settings.json
@@ -1,4 +1,0 @@
-{
-  "accepted_extensions": [".jpg", ".jpeg", ".png", ".gif", ".bmp", ".mp4", ".avi", ".mov", ".mkv"],
-  "language": "en"
-}
