+++ mic_renamer/config/defaults.yaml
@@ -1,0 +1,11 @@
+accepted_extensions:
+  - .jpg
+  - .jpeg
+  - .png
+  - .gif
+  - .bmp
+  - .mp4
+  - .avi
+  - .mov
+  - .mkv
+language: en
