+++ mic_renamer/logic/tag_loader.py
@@ -1,24 +1,24 @@
 import json
 import os
+from pathlib import Path
 
-from ..config.app_config import get_config_dir, load_config
+from .. import config_manager
+from ..utils.path_utils import get_config_dir
 
-DEFAULT_TAGS_FILE = os.path.join(get_config_dir(), "tags.json")
+DEFAULT_TAGS_FILE = Path(get_config_dir()) / "tags.json"
 
-try:
-    CONFIG_TAGS_FILE = load_config().get("tags_file")
-except Exception:
-    CONFIG_TAGS_FILE = None
+CONFIG_TAGS_FILE = config_manager.get("tags_file")
 
 ENV_TAGS_FILE = "RENAMER_TAGS_FILE"
 
 def load_tags(file_path: str | None = None) -> dict:
     """Load tag definitions from a JSON file."""
     if file_path is None:
-        file_path = os.environ.get(ENV_TAGS_FILE, CONFIG_TAGS_FILE or DEFAULT_TAGS_FILE)
-    if not os.path.isfile(file_path):
+        file_path = os.environ.get(ENV_TAGS_FILE) or str(CONFIG_TAGS_FILE or DEFAULT_TAGS_FILE)
+    path = Path(file_path)
+    if not path.is_file():
         return {}
-    with open(file_path, "r", encoding="utf-8") as f:
+    with path.open("r", encoding="utf-8") as f:
         return json.load(f)
