+++ mic_renamer/ui/main_window.py
@@ -3,302 +3,27 @@ import re
 from PySide6.QtWidgets import (
     QWidget, QSplitter, QHBoxLayout, QVBoxLayout, QGridLayout,
     QPushButton, QSlider, QFileDialog, QMessageBox, QToolBar,
-    QGraphicsView, QGraphicsScene, QStyle,
     QApplication, QLabel, QLineEdit,
-    QCheckBox, QTableWidget, QTableWidgetItem, QHeaderView,
-    QProgressDialog, QDialog, QDialogButtonBox, QAbstractItemView
+    QProgressDialog, QDialog, QDialogButtonBox, QAbstractItemView,
+    QHeaderView, QStyle
 )
-from PySide6.QtGui import QPixmap, QPainter, QColor, QAction, QIcon
+from PySide6.QtGui import QColor, QAction, QIcon
 from PySide6.QtCore import Qt, QTimer, QItemSelectionModel, QItemSelection
 
-from .. import config, save_config
-from ..config.app_config import load_config
+from .. import config_manager
 from ..utils.i18n import tr, set_language
 from .settings_dialog import SettingsDialog
+from .panels import ImageViewer, AspectRatioWidget, DragDropTableWidget, TagPanel
 from ..logic.settings import ItemSettings
 from ..logic.renamer import Renamer
-from ..logic.tag_loader import load_tags
+
 
 ROLE_SETTINGS = Qt.UserRole + 1
 
-class ImageViewer(QGraphicsView):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        scene = QGraphicsScene(self)
-        self.setScene(scene)
-        self.pixmap_item = None
-        self.current_pixmap = None
-        # Drag-Pan aktivieren
-        self.setDragMode(QGraphicsView.ScrollHandDrag)
-        # Zoom-Anker unter Maus
-        self.setTransformationAnchor(QGraphicsView.AnchorUnderMouse)
-        self.setResizeAnchor(QGraphicsView.AnchorUnderMouse)
-        # Glättung beim Skalieren
-        self.setRenderHint(QPainter.SmoothPixmapTransform)
-        # Zoom und Rotation intern tracken
-        self._zoom_pct = 100
-        self._rotation = 0  # in Grad, 0/90/180/270 etc.
-        # Focus, damit wheelEvent nur wirkt, wenn Maus über ImageViewer ist
-        self.setFocusPolicy(Qt.StrongFocus)
-
-    def load_image(self, path: str):
-        if not path:
-            self.scene().clear()
-            self.pixmap_item = None
-            self.current_pixmap = None
-            # Reset intern, aber behalte keine Rotation für neuen Bild-Load:
-            self._zoom_pct = 100
-            self._rotation = 0
-            self.reset_transform()
-            return
-        pix = QPixmap(path)
-        if pix.isNull():
-            self.scene().clear()
-            self.pixmap_item = None
-            self.current_pixmap = None
-            self._zoom_pct = 100
-            self._rotation = 0
-            self.reset_transform()
-            return
-        self.current_pixmap = pix
-        self.scene().clear()
-        self.pixmap_item = self.scene().addPixmap(pix)
-        self.scene().setSceneRect(self.pixmap_item.boundingRect())
-        # Nach Laden: initial Fit und keine Rotation
-        self._rotation = 0
-        self.zoom_fit()
-
-    def reset_transform(self):
-        # Rücksetzen der Transformation und Scrollbars an Anfang
-        self.resetTransform()
-        # Setze Scrollbars zurück
-        try:
-            self.horizontalScrollBar().setValue(0)
-            self.verticalScrollBar().setValue(0)
-        except Exception:
-            pass
-
-    def wheelEvent(self, event):
-        # Zoom nur bei Maus über ImageViewer
-        if not self.pixmap_item:
-            return
-        delta = event.angleDelta().y()
-        if delta > 0:
-            self._zoom_pct = min(self._zoom_pct + 10, 500)
-        else:
-            self._zoom_pct = max(self._zoom_pct - 10, 10)
-        self.apply_transformations()
-        event.accept()
-
-    def apply_transformations(self):
-        if not self.pixmap_item:
-            return
-        # Reset
-        self.resetTransform()
-        # Zuerst Rotation, dann Skalierung
-        if self._rotation != 0:
-            # rotate(angle) rotiert relativ; hier nach resetTransform also absolut ok
-            self.rotate(self._rotation)
-        factor = self._zoom_pct / 100.0
-        self.scale(factor, factor)
-        # Scrollbars zurücksetzen, damit obere linke Ecke gezeigt bleibt
-        try:
-            self.horizontalScrollBar().setValue(0)
-            self.verticalScrollBar().setValue(0)
-        except Exception:
-            pass
-
-    def zoom_fit(self):
-        if not self.pixmap_item:
-            return
-        view_rect = self.viewport().rect()
-        scene_rect = self.scene().sceneRect().toRect()
-        if scene_rect.isEmpty():
-            return
-        factor_w = view_rect.width() / scene_rect.width()
-        factor_h = view_rect.height() / scene_rect.height()
-        factor = min(factor_w, factor_h)
-        self._zoom_pct = int(factor * 100)
-        self.resetTransform()
-        if self._rotation != 0:
-            self.rotate(self._rotation)
-        self.scale(factor, factor)
-        try:
-            self.horizontalScrollBar().setValue(0)
-            self.verticalScrollBar().setValue(0)
-        except Exception:
-            pass
-
-    def rotate_left(self):
-        if not self.pixmap_item:
-            return
-        # Update interne Rotation
-        self._rotation = (self._rotation - 90) % 360
-        self.apply_transformations()
-
-    def rotate_right(self):
-        if not self.pixmap_item:
-            return
-        self._rotation = (self._rotation + 90) % 360
-        self.apply_transformations()
-
-class AspectRatioWidget(QWidget):
-    def __init__(self, aspect_ratio=16/9, parent=None):
-        super().__init__(parent)
-        self.aspect_ratio = aspect_ratio
-        self._widget = None
-
-    def setWidget(self, widget):
-        self._widget = widget
-        widget.setParent(self)
-
-    def resizeEvent(self, event):
-        if not self._widget:
-            return super().resizeEvent(event)
-        w = self.width()
-        h = self.height()
-        target_w = w
-        target_h = int(target_w / self.aspect_ratio)
-        if target_h > h:
-            target_h = h
-            target_w = int(target_h * self.aspect_ratio)
-        x = (w - target_w) // 2
-        y = (h - target_h) // 2
-        self._widget.setGeometry(x, y, target_w, target_h)
-        super().resizeEvent(event)
-
-class DragDropTableWidget(QTableWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._updating_checks = False
-        self.setColumnCount(4)
-        self.setHorizontalHeaderLabels(["", "Filename", "Tags", "Suffix"])
-        header = self.horizontalHeader()
-        header.setSectionResizeMode(0, QHeaderView.ResizeToContents)
-        header.setSectionResizeMode(1, QHeaderView.Interactive)
-        header.setSectionResizeMode(2, QHeaderView.Interactive)
-        header.setSectionResizeMode(3, QHeaderView.Stretch)
-        header.setStretchLastSection(True)
-        header.sectionDoubleClicked.connect(self.on_header_double_clicked)
-        self.setHorizontalScrollMode(QAbstractItemView.ScrollPerPixel)
-        self.setAcceptDrops(True)
-        self.setSelectionBehavior(QTableWidget.SelectRows)
-        self.setSelectionMode(QTableWidget.ExtendedSelection)
-        self.verticalHeader().setSectionResizeMode(QHeaderView.Fixed)
-        self.verticalHeader().setDefaultSectionSize(24)
-        self.itemSelectionChanged.connect(self.sync_check_column)
-        self.itemChanged.connect(self.handle_item_changed)
-        self._initial_columns = False
-        QTimer.singleShot(0, self.set_equal_column_widths)
-
-    def set_equal_column_widths(self):
-        if self._initial_columns:
-            return
-        self._initial_columns = True
-        header = self.horizontalHeader()
-        total = self.viewport().width() - header.sectionSize(0)
-        if total <= 0:
-            return
-        w = total // 3
-        for i in range(1, 4):
-            self.setColumnWidth(i, w)
-
-    def on_header_double_clicked(self, index: int):
-        header = self.horizontalHeader()
-        new_width = header.sizeHintForColumn(index)
-        header.resizeSection(index, new_width)
-
-    def dragEnterEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-        else:
-            super().dragEnterEvent(event)
-
-    def dragMoveEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-        else:
-            super().dragMoveEvent(event)
-
-    def dropEvent(self, event):
-        if event.mimeData().hasUrls():
-            paths = []
-            for url in event.mimeData().urls():
-                path = url.toLocalFile()
-                if os.path.isfile(path):
-                    ext = os.path.splitext(path)[1].lower()
-                    if ext in ItemSettings.ACCEPT_EXTENSIONS:
-                        paths.append(path)
-            self.add_paths(paths)
-            event.acceptProposedAction()
-        else:
-            super().dropEvent(event)
-
-    def add_paths(self, paths: list[str]):
-        for path in paths:
-            duplicate = False
-            for row in range(self.rowCount()):
-                item = self.item(row, 1)
-                if item and item.data(Qt.UserRole) == path:
-                    duplicate = True
-                    break
-            if duplicate:
-                continue
-            row = self.rowCount()
-            self.insertRow(row)
-            check_item = QTableWidgetItem()
-            check_item.setFlags(Qt.ItemIsUserCheckable | Qt.ItemIsEnabled)
-            check_item.setCheckState(Qt.Unchecked)
-            fname_item = QTableWidgetItem(os.path.basename(path))
-            fname_item.setData(Qt.UserRole, path)
-            fname_item.setBackground(QColor(30, 30, 30))
-            fname_item.setForeground(QColor(220, 220, 220))
-            tags_item = QTableWidgetItem("")
-            suffix_item = QTableWidgetItem("")
-            tags_item.setToolTip("")
-            suffix_item.setToolTip("")
-            self.setItem(row, 0, check_item)
-            self.setItem(row, 1, fname_item)
-            self.setItem(row, 2, tags_item)
-            self.setItem(row, 3, suffix_item)
-        if self.rowCount() > 0 and not self.selectionModel().hasSelection():
-            self.selectRow(0)
-
-    def sync_check_column(self):
-        selected = {idx.row() for idx in self.selectionModel().selectedRows()}
-        self._updating_checks = True
-        for row in range(self.rowCount()):
-            item = self.item(row, 0)
-            if not item:
-                continue
-            item.setCheckState(Qt.Checked if row in selected else Qt.Unchecked)
-        self._updating_checks = False
-
-    def handle_item_changed(self, item: QTableWidgetItem):
-        if self._updating_checks or item.column() != 0:
-            return
-        row = item.row()
-        index = self.model().index(row, 0)
-        mods = QApplication.keyboardModifiers()
-        if mods & Qt.ShiftModifier and self.selectionModel().hasSelection():
-            cur = self.selectionModel().currentIndex().row()
-            start = min(cur, row)
-            end = max(cur, row)
-            selection = QItemSelection(
-                self.model().index(start, 0),
-                self.model().index(end, self.columnCount() - 1),
-            )
-            command = QItemSelectionModel.Select if item.checkState() == Qt.Checked else QItemSelectionModel.Deselect
-            self.selectionModel().select(selection, command | QItemSelectionModel.Rows)
-        else:
-            if item.checkState() == Qt.Checked:
-                self.selectionModel().select(index, QItemSelectionModel.Select | QItemSelectionModel.Rows)
-            else:
-                self.selectionModel().select(index, QItemSelectionModel.Deselect | QItemSelectionModel.Rows)
-
 class RenamerApp(QWidget):
-    def __init__(self):
+    def __init__(self, state_manager=None):
         super().__init__()
+        self.state_manager = state_manager
         self.setWindowTitle(tr("app_title"))
 
         main_layout = QVBoxLayout(self)
@@ -355,23 +80,12 @@ class RenamerApp(QWidget):
         grid.addWidget(self.table_widget, 0, 1)
 
         # Tag container spanning both columns with manual toggle
-        self.tag_container = QWidget()
-        tag_main_layout = QVBoxLayout(self.tag_container)
-        lbl_tags = QLabel(tr("select_tags_label"))
-        tag_main_layout.addWidget(lbl_tags)
-        self.checkbox_container = QWidget()
-        self.tag_layout = QGridLayout(self.checkbox_container)
-        self.tag_layout.setContentsMargins(0, 0, 0, 0)
-        self.tags_info = {}
-        self.checkbox_map = {}
-        self.rebuild_tag_checkboxes()
-        tag_main_layout.addWidget(self.checkbox_container)
-
+        self.tag_panel = TagPanel()
         self.btn_toggle_tags = QPushButton()
         self.btn_toggle_tags.clicked.connect(self.toggle_tag_panel)
         grid.addWidget(self.btn_toggle_tags, 1, 0, 1, 2, Qt.AlignLeft)
-        grid.addWidget(self.tag_container, 2, 0, 1, 2)
-        self.tag_container.setVisible(False)
+        grid.addWidget(self.tag_panel, 2, 0, 1, 2)
+        self.tag_panel.setVisible(False)
         
         # Initial deaktivieren
         self.set_item_controls_enabled(False)
@@ -380,30 +94,12 @@ class RenamerApp(QWidget):
         self.update_translations()
 
     def toggle_tag_panel(self):
-        visible = self.tag_container.isVisible()
-        self.tag_container.setVisible(not visible)
+        visible = self.tag_panel.isVisible()
+        self.tag_panel.setVisible(not visible)
         self.btn_toggle_tags.setText(tr("hide_tags") if not visible else tr("show_tags"))
 
     def rebuild_tag_checkboxes(self):
-        while self.tag_layout.count():
-            item = self.tag_layout.takeAt(0)
-            w = item.widget()
-            if w:
-                w.deleteLater()
-        self.checkbox_map = {}
-        self.tags_info = load_tags()
-        columns = 4
-        row = col = 0
-        for code, desc in self.tags_info.items():
-            cb = QCheckBox(f"{code}: {desc}")
-            cb.setProperty("code", code)
-            cb.stateChanged.connect(self.save_current_item_settings)
-            self.tag_layout.addWidget(cb, row, col)
-            self.checkbox_map[code] = cb
-            col += 1
-            if col >= columns:
-                col = 0
-                row += 1
+        self.tag_panel.rebuild()
 
     def setup_toolbar(self):
         style = QApplication.style()
@@ -454,7 +150,7 @@ class RenamerApp(QWidget):
     def open_settings(self):
         dlg = SettingsDialog(self)
         if dlg.exec() == QDialog.Accepted:
-            cfg = load_config()
+            cfg = config_manager.load()
             set_language(cfg.get("language", "en"))
             self.update_translations()
             self.rebuild_tag_checkboxes()
@@ -472,7 +168,7 @@ class RenamerApp(QWidget):
         # update form labels
         self.lbl_project.setText(tr("project_number_label"))
         self.input_project.setPlaceholderText(tr("project_number_placeholder"))
-        if self.tag_container.isVisible():
+        if self.tag_panel.isVisible():
             self.btn_toggle_tags.setText(tr("hide_tags"))
         else:
             self.btn_toggle_tags.setText(tr("show_tags"))
@@ -525,7 +221,7 @@ class RenamerApp(QWidget):
         for st in settings_list[1:]:
             intersect &= st.tags
             union |= st.tags
-        for code, cb in self.checkbox_map.items():
+        for code, cb in self.tag_panel.checkbox_map.items():
             cb.blockSignals(True)
             if code in intersect:
                 cb.setTristate(False)
@@ -545,7 +241,7 @@ class RenamerApp(QWidget):
         rows = [idx.row() for idx in self.table_widget.selectionModel().selectedRows()]
         if not rows:
             return
-        checkbox_states = {code: cb.checkState() for code, cb in self.checkbox_map.items()}
+        checkbox_states = {code: cb.checkState() for code, cb in self.tag_panel.checkbox_map.items()}
         for row in rows:
             item0 = self.table_widget.item(row, 1)
             settings: ItemSettings = item0.data(ROLE_SETTINGS)
@@ -591,7 +287,7 @@ class RenamerApp(QWidget):
             return
         if col == 2:
             raw_tags = {t.strip() for t in item.text().split(',') if t.strip()}
-            valid_tags = {t for t in raw_tags if t in self.tags_info}
+            valid_tags = {t for t in raw_tags if t in self.tag_panel.tags_info}
             invalid = raw_tags - valid_tags
             if invalid:
                 QMessageBox.warning(self, "Invalid Tags", 
@@ -636,14 +332,14 @@ class RenamerApp(QWidget):
             self.table_widget.selectRow(row + 1)
 
     def set_item_controls_enabled(self, enabled: bool):
-        for cb in self.checkbox_map.values():
+        for cb in self.tag_panel.checkbox_map.values():
             cb.setEnabled(enabled)
 
     def clear_all(self):
         self.table_widget.setRowCount(0)
         self.image_viewer.load_image("")
         self.zoom_slider.setValue(100)
-        for cb in self.checkbox_map.values():
+        for cb in self.tag_panel.checkbox_map.values():
             cb.setChecked(False)
         self.set_item_controls_enabled(False)
 
@@ -767,4 +463,11 @@ class RenamerApp(QWidget):
             )
         else:
             QMessageBox.information(self, tr("done"), tr("rename_done"))
+
+    def closeEvent(self, event):
+        if self.state_manager:
+            self.state_manager.set("width", self.width())
+            self.state_manager.set("height", self.height())
+            self.state_manager.save()
+        super().closeEvent(event)
