+++ mic_renamer/utils/__init__.py
@@ -1,2 +1,6 @@
 """Utility helpers."""
 
+from .state_manager import StateManager
+from .path_utils import get_config_dir
+
+__all__ = ["StateManager", "get_config_dir"]
