+++ requirements.txt
@@ -1,2 +1,3 @@
 PySide6>=6.0.0
 appdirs>=1.4
+PyYAML>=5.3
